get rid of circular dependency
