+++ packages/atlas-meta-server/src/types.ts
@@ -1,0 +1,13 @@
+export type SchemaOrgTag = {
+  type: 'link' | 'regular'
+  prop: string
+  value: unknown
+}
+
+export type MetaTags = Record<string, string | number>
+
+export type AppData = {
+  name: string
+  orionUrl: string
+  twitterId?: string
+}
